# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Lablink Operational Asset Tracker (OAT) API
  version: 0.1.0
  description: |
    The operational layer for Lablink's laboratory assets: location, custodian, status,
    idle time and utilisation (RFP 2.6.4).

    **The SAP boundary.** SAP S/4HANA FI-AA is the authoritative FINANCIAL record of every
    asset. This API is the OPERATIONAL record. The two are linked by a shared asset ID
    (`sapAssetNo`) and are never merged.

      * `POST /sap/sync` pulls the SAP asset master INTO the OAT (one-way).
      * Only accounting-relevant events (disposal, impairment, cost-centre move) may ever be
        written back to SAP, behind an approval step.
      * Telemetry — heartbeat, idle, utilisation, location — is NEVER sent to SAP. This is
        enforced by the type system, not by convention. See
        `docs/decisions/0004-sap-boundary-typed-contract.md`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000/api
    description: Local development
security:
  - sessionAuth: []
...
paths:
...
  /signals/scan:
    post:
      tags: [Signals]
      summary: Submit a barcode/QR scan
      description: |
        The fallback floor. With every automated connector disabled, this endpoint plus
        manual entry keeps the register fully usable — which is the graceful-degradation
        requirement.

        A scan is an assertion by a human standing in front of the asset, so it is the most
        authoritative signal available and overrides conflicting telemetry.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScanPayload' }
      responses:
        '200':
          description: Scan accepted and the asset re-projected.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestResult' }
        '400': { description: Malformed scan. }
        '401': { $ref: '#/components/responses/Unauthorised' }
        '404': { description: The tag is not in the register. }
        '503': { description: The scan connector is disabled. }
  /connectors/soti/poll:
    post:
      tags: [Signals]
      security:
        - serviceToken: []
      summary: Poll the SOTI MDM and ingest device reports
      description: |
        Feature-flagged by `OAT_CONNECTOR_SOTI`; returns 503 when disabled, which is a
        supported configuration rather than an error.

        Currently backed by a mock (assumption A2). The mock accepts `reports` in the body
        so demos and tests are deterministic; the real adapter fetches from the MDM and
        ignores the body.

        Note the class rules of ADR-0008: a SOTI report can evidence activity for an IT or
        SCANNER asset, but NOT for a LAB_INSTRUMENT, whose activity comes only from the LIS.
        For an instrument this records presence and nothing more.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reports:
                  type: array
                  items: { $ref: '#/components/schemas/SotiDeviceReport' }
      responses:
        '200':
          description: Reports ingested.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunResult' }
        '401': { $ref: '#/components/responses/Unauthorised' }
        '503': { description: The SOTI connector is disabled. }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: |
        Auth.js session cookie, obtained by signing in at `/signin`. Every route additionally
        checks a PERMISSION and a SITE SCOPE — being authenticated is not being authorised.

        Roles per RFP Appendix F: FINANCE, PURCHASING, BRANCH, HQ_LAB_MANAGER, IT, DEVELOPER.
        BRANCH users are scoped to their own site; the scope narrows the query, so it cannot
        be widened by passing another site's id.

    serviceToken:
      type: http
      scheme: bearer
      description: |
        For MACHINE callers only — the scheduler's idle sweep and connector polls, which have
        no session and no person behind them. `OAT_SERVICE_TOKEN`; fails closed (503) when
        unset. A human session does NOT authorise these endpoints, and this token grants
        nothing else.
  responses:
    Unauthorised:
      description: Not signed in, or an invalid/revoked session or service token.
...
  schemas:
    AssetStatus:
      type: string
      enum: [IN_USE, IDLE, UNDER_REPAIR, RETIRED]
      description: |
        `UNDER_REPAIR` and `RETIRED` are administrative: asserted by a human, and never
        overwritten by telemetry.
...
    ScanPayload:
      type: object
      required: [tag]
      properties:
        tag: { type: string }
        observedAt:
          type: string
          format: date-time
          description: Defaults to now. Supply it when submitting scans buffered offline.
        location: { type: string }
        status: { $ref: '#/components/schemas/AssetStatus' }
        custodianId: { type: string }
        scannedBy: { type: string }
    SotiDeviceReport:
      type: object
      required: [deviceId, assetRef, idleMinutes, reportedAt]
      properties:
        deviceId: { type: string }
        assetRef:
          type: string
          description: Asset tag or SAP asset number, per the MDM's custom field.
        idleMinutes:
          type: number
          minimum: 0
          description: |
            Minutes idle at the time of the report. Load-bearing: it dates the idle run from
            when the device went quiet, so an MDM flushing a backlog after an outage is not
            misread as fresh idleness.
        batteryPct: { type: number }
        location: { type: string }
        reportedAt: { type: string, format: date-time }
    IngestResult:
      type: object
      properties:
        accepted:
          type: integer
          description: Signals written.
        duplicates:
          type: integer
          description: Redeliveries suppressed on (source, dedupeKey).
        assetsUpdated:
          type: array
          items: { type: string }
    RunResult:
      allOf:
        - $ref: '#/components/schemas/IngestResult'
        - type: object
          properties:
            connector: { type: string }
            unmatched:
              type: array
              items: { type: string }
              description: |
                External references the connector reported that match no asset. Reported
                rather than auto-created: a misconfigured connector must not be able to
                populate the register.
...